+++ src/GaugeDotnet/Gauges/Models/BaseGaugeSettings.cs
@@ -2,9 +2,9 @@
 {
     public class BaseGaugeSettings
     {
-        public decimal InitialValue { get; set; } = 0;
-        public decimal MinValue { get; set; } = 0;
-        public decimal MaxValue { get; set; } = 100;
+        public float InitialValue { get; set; } = 0;
+        public float MinValue { get; set; } = 0;
+        public float MaxValue { get; set; } = 100;
         public string Unit { get; set; } = "";
         public string Title { get; set; } = "";
     }
